refactor: move `delay` and `highlightElement` to `helper/ui` module
