+++ module/documents/chat.mjs
@@ -1,4 +1,4 @@
-import { highlightElement } from "../utils.mjs";
+import { highlightElement } from "../helper/ui.mjs";
 
 export default class ChatMessageEd extends ChatMessage {
